+++ Gw2Sharp.Tests/WebApi/Http/HttpResponseInfoTests.cs
@@ -1,6 +1,8 @@
 using System;
 using System.Collections.Generic;
 using System.Net;
+using FluentAssertions;
+using FluentAssertions.Extensions;
 using Gw2Sharp.WebApi.Http;
 using Xunit;
 
@@ -13,12 +15,24 @@ namespace Gw2Sharp.Tests.WebApi.Http
         {
             var responseHeaders = new Dictionary<string, string>
             {
-                { "Cache-Control", "\"public, max-age=60\"" }
+                ["Cache-Control"] = "\"public, max-age=60\""
             };
 
             var response = new HttpResponseInfo(HttpStatusCode.OK, responseHeaders);
-            Assert.Equal(responseHeaders, responseHeaders);
-            Assert.Equal(TimeSpan.FromSeconds(60), response.CacheMaxAge);
+            response.RawResponseHeaders.Should().BeEquivalentTo(responseHeaders);
+            response.CacheMaxAge.Should().Be(60.Seconds());
+        }
+
+        [Fact]
+        public void ReadsCustomGw2SharpHeadersTest()
+        {
+            var responseHeaders = new Dictionary<string, string>
+            {
+                ["X-Gw2Sharp-Cache-State"] = CacheState.PartiallyFromCache.ToString()
+            };
+
+            var response = new HttpResponseInfo(HttpStatusCode.OK, responseHeaders);
+            response.CacheState.Should().Be(CacheState.PartiallyFromCache);
         }
     }
 }