+++ Gw2Sharp.Tests/WebApi/V2/RequestBuilderTests.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.Net;
-using System.Threading;
 using System.Threading.Tasks;
+using FluentAssertions;
 using Gw2Sharp.WebApi.Caching;
 using Gw2Sharp.WebApi.Http;
 using Gw2Sharp.WebApi.V2;
@@ -17,28 +17,22 @@ namespace Gw2Sharp.Tests.WebApi.V2
         public async Task ValidRequestTest()
         {
             var httpClient = Substitute.For<IHttpClient>();
-            var httpResponse = Substitute.For<IWebApiResponse>();
-            httpResponse.Content.Returns("{\"testkey\":\"testvalue\"}");
-            httpResponse.StatusCode.Returns(HttpStatusCode.OK);
-            httpResponse.ResponseHeaders.Returns(new Dictionary<string, string> { { "response", "here" } });
-            httpClient.RequestAsync(Arg.Any<IWebApiRequest>(), CancellationToken.None).Returns(Task.FromResult(httpResponse));
+            var httpResponse = new WebApiResponse(
+                "{\"testkey\":\"testvalue\"}",
+                HttpStatusCode.OK,
+                new Dictionary<string, string> { { "response", "here" } });
+            httpClient.RequestAsync(Arg.Any<IWebApiRequest>()).Returns(_ => httpResponse);
             var content = new TestContentClass { Testkey = "testvalue" };
 
             var connection = new Connection(string.Empty, default, new NullCacheMethod(), httpClient: httpClient);
-            var request = new WebApiRequest(new WebApiRequestOptions
-            {
-                BaseUrl = "http://localhost",
-                EndpointPath = "endpoint",
-                BulkQueryParameterIdName = "id",
-                BulkQueryParameterIdsName = "ids"
-            }, connection, Substitute.For<IGw2Client>());
             var response = await new RequestBuilder(Substitute.For<IEndpointClient>(), connection, Substitute.For<IGw2Client>())
                 .Blob()
                 .ExecuteAsync<TestContentClass>();
 
-            Assert.Equal(content.Testkey, response.Content.Testkey);
-            Assert.Equal(httpResponse.StatusCode, response.StatusCode);
-            Assert.Equal(httpResponse.ResponseHeaders, response.ResponseHeaders);
+            response.Content.Should().BeEquivalentTo(content);
+            response.StatusCode.Should().Be(httpResponse.StatusCode);
+            response.ResponseHeaders.Should().Contain(httpResponse.ResponseHeaders);
+            response.ResponseHeaders.Should().Contain("X-Gw2Sharp-Cache-State", CacheState.FromLive.ToString());
         }